added Ontology and InferenceOptions to rdfbean.model added DefaultOntology with MappedClass inspection updated Guice configuration to use Ontology

+++ rdfbean-sesame/src/test/java/com/mysema/rdfbean/sesame/query/BeanQueryStandardTest.java
@@ -84,7 +84,7 @@ public class BeanQueryStandardTest extends SessionTestBase {
         
     @Override
     public void tearDown() throws IOException{
-        session.delete(other);
+        if (other != null) session.delete(other);
         super.tearDown();
     }
